+++ src/cruiz/commands/context.py
@@ -17,10 +17,10 @@ from cruiz.recipe.logs.command import CommandListWidgetItem, RecipeCommandHistor
 from cruiz.settings.managers.namedlocalcache import NamedLocalCacheSettingsReader
 
 import cruizlib.workers.api as workers_api
+from cruizlib.commands.metarequestconaninvocation import MetaRequestConanInvocation
 from cruizlib.constants import DEFAULT_CACHE_NAME
 from cruizlib.exceptions import RecipeInspectionError
 from cruizlib.interop.commandparameters import CommandParameters
-from cruizlib.workers.metarequestconaninvocation import MetaRequestConanInvocation
 from cruizlib.workers.utils.text2html import text_to_html
 
 from .conanenv import get_conan_env
